--- api/auth/utils.go
@@ -1,21 +1,0 @@
-package auth
-
-import (
-	"golang.org/x/crypto/bcrypt"
-)
-
-func HashPassword(password string) (string, error) {
-	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	if err != nil {
-		return "", err
-	}
-	return string(hashedPassword), nil
-}
-
-func VerifyPassword(hashed, password string) bool {
-	check := bcrypt.CompareHashAndPassword([]byte(hashed), []byte(password))
-	if check != nil {
-		return false
-	}
-	return true
-}
