+++ api/tools/networks/net_test.go
@@ -5,7 +5,7 @@ import (
 
 	"testing"
 
-	"github.com/danmuck/dps_http/api/logs"
+	"github.com/danmuck/dps_http/lib/logs"
 	"github.com/shirou/gopsutil/net"
 )
