+++ lib/middleware/auth.go
@@ -5,7 +5,7 @@ import (
 
 	"slices"
 
-	"github.com/danmuck/dps_http/api/logs"
+	"github.com/danmuck/dps_http/lib/logs"
 	"github.com/gin-gonic/gin"
 	"github.com/golang-jwt/jwt/v4"
 )
